# openapi/organization-service.yaml
openapi: 3.0.3

info:
  title: Organization Service
  version: 1.0.0
  description: |
    Первый веб-сервис управляет коллекцией объектов класса `Organization`.

    Через REST API можно создать организацию, получить её по ИД, обновить,
    удалить или запросить массив элементов коллекции. На отдельных URL доступны
    группировка по `name`, подсчёт объектов с `annualTurnover` меньше заданного
    и получение массива уникальных значений `annualTurnover`.

    Все данные передаются в формате XML (`application/xml`).
    При нарушении ограничений целостности сервис возвращает код,
    соответствующий ошибке (400/404).

servers:
  - url: /

security: []

tags:
  - name: Organizations
    description: Базовые операции над объектами коллекции
  - name: Employees
    description: Управление сотрудниками, связанными с организацией (многие к одному)
  - name: Statistics
    description: Дополнительные операции агрегации и выборки

paths:
  /organizations:
    get:
      tags: [Organizations]
      operationId: getOrganizations
      summary: Получить массив элементов коллекции
      description: |
        Возвращает организации. Можно задать фильтры, сортировку и страницу.
        Сервис сначала применяет фильтры, затем подсчитывает `total` и выбирает
        страницу. `total` содержит число всех организаций после фильтрации.
        Если `page` выходит за пределы результата, сервис возвращает пустой
        `items` с тем же значением `total`.
      parameters:
        - $ref: '#/components/parameters/filterId'
        - $ref: '#/components/parameters/filterName'
        - $ref: '#/components/parameters/filterCoordinatesX'
        - $ref: '#/components/parameters/filterCoordinatesY'
        - $ref: '#/components/parameters/filterCreationDate'
        - $ref: '#/components/parameters/filterAnnualTurnover'
        - $ref: '#/components/parameters/filterFullName'
        - $ref: '#/components/parameters/filterType'
        - $ref: '#/components/parameters/filterStreet'
        - $ref: '#/components/parameters/sortBy'
        - $ref: '#/components/parameters/sortOrder'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: Постраничный результат выборки
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/OrganizationList'
              example: |
                <?xml version="1.0" encoding="UTF-8"?>
                <organizations>
                  <total>1</total>
                  <page>1</page>
                  <size>10</size>
                  <organization id="1">
                    <name>Acme</name>
                    <coordinates>
                      <x>12.5</x>
                      <y>34</y>
                    </coordinates>
                    <creationDate>2024-03-01</creationDate>
                    <annualTurnover>1000000</annualTurnover>
                    <fullName>Acme Corporation</fullName>
                    <type>COMMERCIAL</type>
                    <postalAddress>
                      <street>Main street</street>
                    </postalAddress>
                  </organization>
                </organizations>
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/EndpointNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    post:
      tags: [Organizations]
      operationId: createOrganization
      summary: Добавить новый элемент в коллекцию
      description: |
        Добавляет организацию. `id` и `creationDate` задаются автоматически.
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              $ref: '#/components/schemas/OrganizationInput'
            example: |
              <?xml version="1.0" encoding="UTF-8"?>
              <organization>
                <name>Acme</name>
                <coordinates>
                  <x>12.5</x>
                  <y>34</y>
                </coordinates>
                <annualTurnover>1000000</annualTurnover>
                <fullName>Acme Corporation</fullName>
                <type>COMMERCIAL</type>
                <postalAddress>
                  <street>Main street</street>
                </postalAddress>
              </organization>
      responses:
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '201':
          description: Объект создан
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/EndpointNotFound'

        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /organizations/{id}:
    parameters:
      - $ref: '#/components/parameters/pathId'
    get:
      tags: [Organizations]
      operationId: getOrganizationById
      summary: Получить элемент по ИД
      description: Возвращает организацию с указанным ID.
      responses:
        '200':
          description: Найденный объект
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Organization'
        '404':
          $ref: '#/components/responses/EndpointNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    put:
      tags: [Organizations]
      operationId: updateOrganization
      summary: Обновить элемент
      description: |
        Обновляет организацию. `id` и `creationDate` остаются прежними.
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              $ref: '#/components/schemas/OrganizationInput'
      responses:
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '200':
          description: Обновлённый объект
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/EndpointNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    delete:
      tags: [Organizations]
      operationId: deleteOrganization
      summary: Удалить элемент
      description: Удаляет организацию с указанным ID.
      responses:
        '204':
          description: Объект удалён
        '404':
          $ref: '#/components/responses/EndpointNotFound'

        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /organizations/{id}/employees:
    parameters:
      - $ref: '#/components/parameters/pathId'
    get:
      tags: [Employees]
      operationId: getOrganizationEmployees
      summary: Получить сотрудников организации
      description: |
        Возвращает сотрудников этой организации.
      responses:
        '200':
          description: Список сотрудников
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/EmployeeList'
        '404':
          $ref: '#/components/responses/EndpointNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    post:
      tags: [Employees]
      operationId: addOrganizationEmployee
      summary: Добавить сотрудника в организацию
      description: |
        Добавляет сотрудника в организацию, указанную в пути запроса.
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              $ref: '#/components/schemas/EmployeeInput'
            example: |
              <?xml version="1.0" encoding="UTF-8"?>
              <employee>
                <name>John Doe</name>
                <position>Engineer</position>
                <salary>75000</salary>
              </employee>
      responses:
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '201':
          description: Сотрудник создан
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/EndpointNotFound'

        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /organizations/{id}/employees/{employeeId}:
    parameters:
      - $ref: '#/components/parameters/pathId'
      - $ref: '#/components/parameters/pathEmployeeId'
    delete:
      tags: [Employees]
      operationId: deleteOrganizationEmployee
      summary: Удалить сотрудника организации
      description: Удаляет сотрудника из этой организации.
      responses:
        '204':
          description: Сотрудник удалён
        '404':
          $ref: '#/components/responses/EndpointNotFound'

        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /organizations/stats/grouped-by-name:
    get:
      tags: [Statistics]
      operationId: getOrganizationsGroupedByName
      summary: Сгруппировать объекты по значению поля name
      description: |
        Считает организации с одинаковым `name`.
      responses:
        '200':
          description: Группы и количество элементов в каждой
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/NameGroupList'
              example: |
                <?xml version="1.0" encoding="UTF-8"?>
                <groups>
                  <group>
                    <name>Acme</name>
                    <count>3</count>
                  </group>
                  <group>
                    <name>Globex</name>
                    <count>1</count>
                  </group>
                </groups>
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '404':
          $ref: '#/components/responses/EndpointNotFound'

        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /organizations/stats/annual-turnover/less-than/{value}:
    parameters:
      - name: value
        in: path
        required: true
        description: Пороговое значение поля annualTurnover
        schema:
          type: integer
          format: int32
          minimum: 1
    get:
      tags: [Statistics]
      operationId: countAnnualTurnoverLessThan
      summary: Количество объектов с annualTurnover меньше заданного
      description: |
        Считает организации с `annualTurnover` меньше указанного значения.
      responses:
        '200':
          description: Количество объектов
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/CountResponse'
              example: |
                <?xml version="1.0" encoding="UTF-8"?>
                <count>5</count>
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/EndpointNotFound'

        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /organizations/stats/annual-turnover/unique:
    get:
      tags: [Statistics]
      operationId: getUniqueAnnualTurnover
      summary: Уникальные значения поля annualTurnover
      description: |
        Возвращает уникальные значения `annualTurnover`, включая `null`, если
        хотя бы у одной организации оборот не задан. В XML значение `null`
        представляется элементом `<value xsi:nil="true"/>`.
      responses:
        '200':
          description: Массив уникальных значений
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/UniqueTurnoverList'
              example: |
                <?xml version="1.0" encoding="UTF-8"?>
                <values xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
                  <value>1000000</value>
                  <value>2500000</value>
                  <value xsi:nil="true"/>
                </values>
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '404':
          $ref: '#/components/responses/EndpointNotFound'

        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  parameters:
    pathId:
      name: id
      in: path
      required: true
      description: Идентификатор организации (больше 0)
      schema:
        type: integer
        format: int32
        minimum: 1
    pathEmployeeId:
      name: employeeId
      in: path
      required: true
      description: Идентификатор сотрудника (больше 0)
      schema:
        type: integer
        format: int32
        minimum: 1
    filterId:
      name: id
      in: query
      required: false
      description: Фильтр по полю id
      schema:
        type: integer
        format: int32
        minimum: 1
    filterName:
      name: name
      in: query
      required: false
      description: Фильтр по полю name
      schema:
        type: string
    filterCoordinatesX:
      name: coordinates.x
      in: query
      required: false
      description: Фильтр по полю coordinates.x
      schema:
        type: number
        format: float
    filterCoordinatesY:
      name: coordinates.y
      in: query
      required: false
      description: Фильтр по полю coordinates.y
      schema:
        type: integer
        format: int64
    filterCreationDate:
      name: creationDate
      in: query
      required: false
      description: Фильтр по полю creationDate
      schema:
        type: string
        format: date
    filterAnnualTurnover:
      name: annualTurnover
      in: query
      required: false
      description: Фильтр по полю annualTurnover
      schema:
        type: integer
        format: int32
        minimum: 1
    filterFullName:
      name: fullName
      in: query
      required: false
      description: Фильтр по полю fullName
      schema:
        type: string
    filterType:
      name: type
      in: query
      required: false
      description: Фильтр по полю type
      schema:
        $ref: '#/components/schemas/OrganizationType'
    filterStreet:
      name: postalAddress.street
      in: query
      required: false
      description: Фильтр по полю postalAddress.street
      schema:
        type: string
        maxLength: 158
    sortBy:
      name: sortBy
      in: query
      required: false
      description: |
        Поле сортировки. Чтобы сортировать по нескольким полям, повторите
        параметр в нужном порядке. Допустимы все поля класса, включая вложенные:
        `id`, `name`, `coordinates.x`, `coordinates.y`, `creationDate`,
        `annualTurnover`, `fullName`, `type`, `postalAddress.street`.
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - id
            - name
            - coordinates.x
            - coordinates.y
            - creationDate
            - annualTurnover
            - fullName
            - type
            - postalAddress.street
    sortOrder:
      name: sortOrder
      in: query
      required: false
      description: Порядок сортировки, применяемый к полям из sortBy
      schema:
        type: string
        enum: [asc, desc]
        default: asc
    page:
      name: page
      in: query
      required: false
      description: Номер страницы, начиная с 1. Страница за пределами результата возвращает пустой items.
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 1
    size:
      name: size
      in: query
      required: false
      description: Максимальное число элементов на странице. Сервис выбирает страницу после фильтрации, перед возвратом результата.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 10

  responses:
    BadRequest:
      description: Переданы данные, нарушающие ограничения целостности
      content:
        application/xml:
          schema:
            $ref: '#/components/schemas/Error'
          example: |
            <?xml version="1.0" encoding="UTF-8"?>
            <error>
              <code>400</code>
              <message>Поле name не может быть пустым</message>
            </error>
    UnsupportedMediaType:
      description: Тело запроса должно быть передано в формате application/xml
      content:
        application/xml:
          schema:
            $ref: '#/components/schemas/Error'
          example: |
            <?xml version="1.0" encoding="UTF-8"?>
            <error>
              <code>415</code>
              <message>Ожидался формат application/xml</message>
            </error>
    EndpointNotFound:
      description: Запрошенный путь или объект не найдены
      content:
        application/xml:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            endpointNotFound:
              summary: Запрошенный путь не найден
              value: |
                <?xml version="1.0" encoding="UTF-8"?>
                <error>
                  <code>404</code>
                  <message>Запрошенная ручка не найдена</message>
                </error>
            itemNotFound:
              summary: Объект с указанным идентификатором не найден
              value: |
                <?xml version="1.0" encoding="UTF-8"?>
                <error>
                  <code>404</code>
                  <message>Объект с указанным идентификатором не найден</message>
                </error>
    InternalServerError:
      description: Внутренняя ошибка сервиса, не связанная с временной недоступностью БД
      content:
        application/xml:
          schema:
            $ref: '#/components/schemas/Error'
          example: |
            <?xml version="1.0" encoding="UTF-8"?>
            <error>
              <code>500</code>
              <message>Внутренняя ошибка сервиса</message>
            </error>
    ServiceUnavailable:
      description: Сервис временно не может обработать запрос, например из-за недоступности БД
      content:
        application/xml:
          schema:
            $ref: '#/components/schemas/Error'
          example: |
            <?xml version="1.0" encoding="UTF-8"?>
            <error>
              <code>503</code>
              <message>База данных временно недоступна</message>
            </error>

  schemas:
    Organization:
      type: object
      xml:
        name: organization
      required:
        - id
        - name
        - coordinates
        - creationDate
        - type
        - postalAddress
      properties:
        id:
          type: integer
          format: int32
          minimum: 1
          readOnly: true
          description: Уникальный идентификатор, генерируется автоматически
          xml:
            attribute: true
        name:
          type: string
          minLength: 1
          description: Имя организации. Не допускает null или пустую строку.
        coordinates:
          $ref: '#/components/schemas/Coordinates'
        creationDate:
          type: string
          format: date
          readOnly: true
          description: Дата создания, генерируется автоматически
        annualTurnover:
          type: integer
          format: int32
          minimum: 1
          nullable: true
          description: Годовой оборот. Может быть null; если задан, должен быть больше 0.
        fullName:
          type: string
          nullable: true
          description: Полное наименование. Может быть null.
        type:
          $ref: '#/components/schemas/OrganizationType'
        postalAddress:
          $ref: '#/components/schemas/Address'

    OrganizationInput:
      type: object
      xml:
        name: organization
      required:
        - name
        - coordinates
        - type
        - postalAddress
      properties:
        name:
          type: string
          minLength: 1
          description: Имя организации. Не допускает null или пустую строку.
        coordinates:
          $ref: '#/components/schemas/Coordinates'
        annualTurnover:
          type: integer
          format: int32
          minimum: 1
          nullable: true
          description: Годовой оборот. Может быть null; если задан, должен быть больше 0.
        fullName:
          type: string
          nullable: true
          description: Полное наименование. Может быть null.
        type:
          $ref: '#/components/schemas/OrganizationType'
        postalAddress:
          $ref: '#/components/schemas/Address'

    Coordinates:
      type: object
      xml:
        name: coordinates
      required:
        - x
        - y
      properties:
        x:
          type: number
          format: float
          description: Координата X
        y:
          type: integer
          format: int64
          description: Координата Y

    Address:
      type: object
      xml:
        name: postalAddress
      required:
        - street
      properties:
        street:
          type: string
          maxLength: 158
          description: Улица. Не допускает null; длина не больше 158 символов.

    OrganizationType:
      type: string
      description: Тип организации
      enum:
        - COMMERCIAL
        - TRUST
        - PRIVATE_LIMITED_COMPANY
        - OPEN_JOINT_STOCK_COMPANY

    Employee:
      type: object
      xml:
        name: employee
      required:
        - id
        - name
        - organizationId
      properties:
        id:
          type: integer
          format: int32
          minimum: 1
          readOnly: true
          description: Уникальный идентификатор сотрудника
          xml:
            attribute: true
        name:
          type: string
          minLength: 1
          description: Имя сотрудника
        position:
          type: string
          nullable: true
          description: Должность
        salary:
          type: number
          format: double
          minimum: 0
          nullable: true
          description: Зарплата
        organizationId:
          type: integer
          format: int32
          minimum: 1
          description: ИД организации (связь многие к одному)
          xml:
            attribute: true

    EmployeeInput:
      type: object
      xml:
        name: employee
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          description: Имя сотрудника
        position:
          type: string
          nullable: true
          description: Должность
        salary:
          type: number
          format: double
          minimum: 0
          nullable: true
          description: Зарплата

    OrganizationList:
      type: object
      xml:
        name: organizations
      required:
        - total
        - page
        - size
      properties:
        total:
          type: integer
          format: int64
          description: Число элементов после фильтрации и до пагинации
        page:
          type: integer
          format: int32
          description: Номер текущей страницы, начиная с 1; страница за пределами результата содержит пустой items
        size:
          type: integer
          format: int32
          description: Запрошенный максимальный размер страницы
        items:
          type: array
          xml:
            name: organization
            wrapped: false
          items:
            $ref: '#/components/schemas/Organization'

    EmployeeList:
      type: object
      xml:
        name: employees
      properties:
        items:
          type: array
          xml:
            name: employee
            wrapped: false
          items:
            $ref: '#/components/schemas/Employee'

    NameGroup:
      type: object
      xml:
        name: group
      required:
        - name
        - count
      properties:
        name:
          type: string
          description: Значение поля name
        count:
          type: integer
          format: int64
          description: Количество элементов в группе

    NameGroupList:
      type: object
      xml:
        name: groups
      properties:
        items:
          type: array
          xml:
            name: group
            wrapped: false
          items:
            $ref: '#/components/schemas/NameGroup'

    CountResponse:
      type: object
      xml:
        name: count
      required:
        - count
      properties:
        count:
          type: integer
          format: int64
          description: Количество объектов

    UniqueTurnoverList:
      type: object
      xml:
        name: values
      properties:
        items:
          type: array
          description: Уникальные значения annualTurnover
          xml:
            name: value
            wrapped: false
          items:
            type: integer
            format: int32
            nullable: true
            description: null кодируется XML-элементом value с xsi:nil="true"

    Error:
      type: object
      xml:
        name: error
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: HTTP-код ошибки
        message:
          type: string
          description: Сообщение об ошибке